Architecture reorganized
