+++ db/connectorMethods.go
@@ -1,26 +1,10 @@
-package postgres
+package db
 
 import (
 	"fmt"
 
 	"github.com/jmoiron/sqlx"
-
-	_ "github.com/lib/pq"
 )
-
-var DBConf = struct {
-	host     string
-	port     int
-	user     string
-	password string
-	dbname   string
-}{
-	host:     "localhost",
-	port:     5432,
-	user:     "postgres",
-	password: "postgres",
-	dbname:   "pinterest",
-}
 
 var SQLStatements = map[string]string{
 	"RegisterUser":   `INSERT INTO public.users ("email", "nickname", "password") VALUES($1, $2, $3)`,
@@ -29,7 +13,10 @@ var SQLStatements = map[string]string{
 	"GetPinsOfUser":  `SELECT * FROM public.pins WHERE author_id=$1`,
 }
 
-func getPinsByRows(rows *sqlx.Rows) (*[]Pin, error) {
+// ------------ Users ------------
+
+// ------------ Pins ------------
+func getPinsByRows(rows *sqlx.Rows) ([]Pin, error) {
 	var result []Pin
 	var pin = &Pin{}
 	for rows.Next() {
@@ -39,10 +26,10 @@ func getPinsByRows(rows *sqlx.Rows) (*[]Pin, error) {
 		}
 		result = append(result, *pin)
 	}
-	return &result, nil
+	return result, nil
 }
 
-func (handler *APIHandler) GetAllPins() (*[]Pin, error) {
+func (handler *Connector) GetAllPins() ([]Pin, error) {
 	rows, err := handler.db.Queryx(SQLStatements["GetAllPins"])
 	fmt.Println(rows)
 	if err != nil {
@@ -51,7 +38,7 @@ func (handler *APIHandler) GetAllPins() (*[]Pin, error) {
 	return getPinsByRows(rows)
 }
 
-func (handler *APIHandler) GetPinsOfUser(userId int) (*[]Pin, error) {
+func (handler *Connector) GetPinsOfUser(userId int) ([]Pin, error) {
 	rows, err := handler.db.Queryx(SQLStatements["GetPinsOfUser"], userId)
 	if err != nil {
 		return nil, err
