+++ db/models.go
@@ -1,4 +1,4 @@
-package postgres
+package db
 
 import "time"
 
@@ -9,4 +9,12 @@ type Pin struct {
 	Caption    string    `db:"caption"`
 	ClickUrl   string    `db:"click_url"`
 	ContentUrl string    `db:"content_url"`
+}
+
+type User struct {
+	UserId     int64     `db:"user_id"`
+	Email      string    `db:"email"`
+	Nickname   string    `db:"nickname"`
+	Password   string    `db:"password"`
+	RegisterAt time.Time `db:"register_at"`
 }
